Praparation for Client/Server refactor
