+++ EasySslStream/Connection/Client/SslClient.cs
@@ -1,329 +1,37 @@
-﻿using System.Collections.Concurrent;
-using System.Net;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net.Security;
 using System.Net.Sockets;
+using System.Net;
 using System.Security.Authentication;
 using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Channels;
+using System.Threading.Tasks;
 
 namespace EasySslStream.Connection.Client
 {
-    /// <summary>
-    /// Server class that can handle multiple clients
-    /// </summary>
-    public class Server
-    {
-        /// <summary>
-        /// List that contains connected clients 
-        /// </summary>
-        public List<SSLClient> ConnectedClients = new List<SSLClient>();
-
-        /// <summary>
-        /// Thread safe dictionary that contains connected clients referenced by int
-        /// </summary>
-        public ConcurrentDictionary<int, SSLClient> ConnectedClientsByNumber = new ConcurrentDictionary<int, SSLClient>();
-
-
-        /// <summary>
-        /// Thread safe dictionary that contains connected clients referenced by string endpoint ( 127.0.0.1:5000 etc)
-        /// </summary>
-        public ConcurrentDictionary<IPEndPoint, SSLClient> ConnectedClientsByEndPoint = new ConcurrentDictionary<IPEndPoint, SSLClient>();
-
-
-        private X509Certificate2 serverCert = null;
-
-
-        private TcpListener listener = null;
-
-        private CancellationTokenSource cts = new CancellationTokenSource();
-
-        /// <summary>
-        /// Encoding for text messages
-        /// </summary>
-        public Encoding TextReceiveEncoding = Encoding.UTF8;
-
-        /// <summary>
-        /// Encoding for filenames
-        /// </summary>
-        public Encoding FileNameEncoding = Encoding.UTF8;
-
-        /// <summary>
-        /// Specifies how certificate verification should behave
-        /// </summary>
-        public CertificateCheckSettings CertificateCheckSettings = new CertificateCheckSettings();
-
-        /// <summary>
-        /// Action Delegate for handling text data received from client, by default it prints message by Console.WriteLine()
-        /// </summary>
-        public Action<string> HandleReceivedText = (string text) =>
-        {
-            Console.WriteLine(text);
-        };
-
-        /// <summary>
-        /// Action Delegate for handling bytes received from client, by default it prints int representation of them in console
-        /// </summary>
-        public Action<byte[]> HandleReceivedBytes = (byte[] bytes) =>
-        {
-            foreach (byte b in bytes) { Console.Write(Convert.ToInt32(b) + " "); }
-            //return bytes
-        };
-
-        // Shutting down the server
-        ////////////////////////////////////////////////////////////////////  
-        private TaskCompletionSource<object> GentleStopLock = new TaskCompletionSource<object>();
-        internal void WorkLock()
-        {
-            bool NoJobs = true;
-            foreach (SSLClient client in ConnectedClients)
-            {
-                if (client.Busy)
-                {
-                    NoJobs = false;
-                }
-            }
-
-            if (NoJobs == true)
-            {
-                if (!GentleStopLock.Task.IsCompleted)
-                    GentleStopLock.SetResult(null);
-            }
-
-
-
-        }
-
-        /// <summary>
-        /// Waits for currently running transfers to end, for all connections, then shuts down the server.
-        /// </summary>
-        public async Task GentleStopServer(int interval = 100)
-        {
-            if (ConnectedClients.Count != 0)
-            {
-                Console.WriteLine("Waiting for all jobs to terminate");
-                bool loopcancel = true;
-                Task.Run(() =>
-                {
-                    while (loopcancel)
-                    {
-                        WorkLock();
-                        Task.Delay(interval).Wait();
-                    }
-
-                });
-
-
-
-                await GentleStopLock.Task;
-
-                loopcancel = false;
-
-                Parallel.ForEach(ConnectedClients, SSLClient =>
-                {
-                    SSLClient.Stop();
-                });
-                listener.Stop();
-
-
-            }
-            else
-            {
-                Parallel.ForEach(ConnectedClients, SSLClient =>
-                {
-                    SSLClient.Stop();
-                });
-                listener.Stop();
-            }
-
-        }
-
-
-
-
-        /// <summary>
-        /// Disposes all connected clients and stops server from listening
-        /// </summary>
-        public void StopServer()
-        {
-            Parallel.ForEach(ConnectedClients, SSLClient =>
-            {
-                SSLClient.Stop();
-            });
-
-            this.listener.Stop();
-        }
-
-
-        /// <summary>
-        /// Sends text Message to client
-        /// </summary>
-        /// <param name="clientEndpoint">Client endpoint</param>
-        /// <param name="Message">byte array representation of the message</param>
-        public void WriteTextToClient(IPEndPoint clientEndpoint, byte[] Message)
-        {
-            ConnectedClientsByEndPoint[clientEndpoint].WriteText(Message);
-        }
-
-        /// <summary>
-        /// Sends text Message to client
-        /// </summary>
-        /// <param name="ConnectionID"></param>
-        /// <param name="Message"></param>
-        public void WriteTextToClient(int ConnectionID, byte[] Message)
-        {
-            ConnectedClients[ConnectionID].WriteText(Message);
-        }
-
-        public void SendRawBytesToClient(int ConnectionID, byte[] Message)
-        {
-            ConnectedClients[ConnectionID].SendRawBytes(Message);
-        }
-
-        public void SendRawBytesToClient(IPEndPoint clientEndpoint, byte[] Message)
-        {
-            ConnectedClientsByEndPoint[clientEndpoint].SendRawBytes(Message);
-        }
-
-        /// <summary>
-        /// Sends file to client
-        /// </summary>
-        /// <param name="ConnectionID">Id of connection</param>
-        /// <param name="Path">path to the file to send</param>
-        public void WriteFileToClient(int ConnectionID, string Path)
-        {
-            ConnectedClients[ConnectionID].SendFile(Path);
-        }
-
-        /// <summary>
-        /// Sends file to client
-        /// </summary>
-        /// <param name="clientEndpoint">client endpoint</param>
-        /// <param name="Path">path to the file to send</param>
-        public void WriteFileToClient(IPEndPoint clientEndpoint, string Path)
-        {
-            ConnectedClientsByEndPoint[clientEndpoint].SendFile(Path);
-        }
-
-        /// <summary>
-        /// Sends Directory to client
-        /// </summary>
-        /// <param name="ConnectionID">Connection id</param>
-        /// <param name="Path">Path to directory to send</param>
-        /// <param name="StopAndThrowOnFailedTransfer">Stops transfer ic coulndn't read the file,if true.If false ignores any errors</param>
-        /// <param name="FailSafeInterval">If connection crashes try to raise this value</param>
-        public void SendDirectoryToClient(int ConnectionID, string Path, bool StopAndThrowOnFailedTransfer = true, int FailSafeInterval = 20)
-        {
-            ConnectedClientsByNumber[ConnectionID].SendDirectory(Path, StopAndThrowOnFailedTransfer, FailSafeInterval);
-        }
-
-        /// <summary>
-        /// Sends Directory to client
-        /// </summary>
-        /// <param name="clientEndPoint">Client endpoint</param>
-        /// <param name="Path">Path to directory to send</param>
-        /// <param name="StopAndThrowOnFailedTransfer">Stops transfer ic coulndn't read the file,if true.If false ignores any errors</param>
-        /// <param name="FailSafeInterval">If connection crashes try to raise this value</param>
-        public void SendDirectoryToClient(IPEndPoint clientEndPoint, string Path, bool StopAndThrowOnFailedTransfer = true, int FailSafeInterval = 20)
-        {
-            ConnectedClientsByEndPoint[clientEndPoint].SendDirectory(Path, StopAndThrowOnFailedTransfer, FailSafeInterval);
-        }
-
-        /// <summary>
-        /// Optimized - Sends directory to client
-        /// </summary>
-        /// <param name="ConnectionID"></param>
-        /// <param name="Path"></param>
-        /// <param name="StopAndThrowOnFailedTransfer"></param>
-        /// <param name="FailSafeInterval"></param>
-        public void SendDirectoryToClientV2(int ConnectionID, string Path, bool StopAndThrowOnFailedTransfer = true, int FailSafeInterval = 20)
-        {
-            ConnectedClientsByNumber[ConnectionID].SendDirectoryV2(Path, StopAndThrowOnFailedTransfer, FailSafeInterval);
-        }
-
-        public void SendDirectoryToClientV2(IPEndPoint clientEndPoint, string Path, bool StopAndThrowOnFailedTransfer = true, int FailSafeInterval = 20)
-        {
-            ConnectedClientsByEndPoint[clientEndPoint].SendDirectoryV2(Path, StopAndThrowOnFailedTransfer, FailSafeInterval);
-        }
-
-
-
-
-        /// <summary>
-        /// Location for the received file from clients
-        /// </summary>
-        public string ReceivedFilesLocation = AppDomain.CurrentDomain.BaseDirectory;
-
-
-        /// <summary>
-        /// Starts server
-        /// </summary>
-        /// <param name="ListenOnIp">Listening ip</param>
-        /// <param name="port">Listening port</param>
-        /// <param name="ServerPFXCertificatePath">Path to the Certificate with private key in pfx format</param>
-        /// <param name="CertPassword">Password to the certificate use empty string if there's no password</param>
-        /// <param name="VerifyClients">Set true if server is meant to check for client certificate, otherwise set false</param>
-        public void StartServer(string ListenOnIp, int port, string ServerPFXCertificatePath, string CertPassword, bool VerifyClients)
-        {
-
-            serverCert = new X509Certificate2(ServerPFXCertificatePath, CertPassword, X509KeyStorageFlags.PersistKeySet);
-            listener = new TcpListener(IPAddress.Parse(ListenOnIp), port);
-            Thread listenerThread = new Thread(() =>
-            {
-                listener.Start();
-                int connected = 0;
-                while (listener.Server.IsBound)
-                {
-                    TcpClient client = listener.AcceptTcpClient();
-                    SSLClient connection = new SSLClient(client, serverCert, VerifyClients, this);
-                    Console.WriteLine(client.Client.RemoteEndPoint?.ToString());
-                }
-                connected++;
-            });
-            listenerThread.Start();
-        }
-
-        /// <summary>
-        /// Starts server
-        /// </summary>
-        /// <param name="ListenOnIp">Listening ip</param>
-        /// <param name="port">Listening port</param>
-        /// <param name="ServerPFXCertificatePath">Path to the Certificate with private key in pfx format</param>
-        /// <param name="CertPassword">Password to the certificate use empty string if there's no password</param>
-        /// <param name="VerifyClients">Set true if server is meant to check for client certificate, otherwise set false</param>
-        public void StartServer(IPAddress ListenOnIp, int port, string ServerPFXCertificatePath, string CertPassword, bool VerifyClients)
-        {
-
-            this.serverCert = new X509Certificate2(ServerPFXCertificatePath, CertPassword, X509KeyStorageFlags.PersistKeySet);
-            listener = new TcpListener(ListenOnIp, port);
-            listener.Start();
-            Thread listenerThrewad = new Thread(() =>
-            {
-                int connected = 0;
-                while (listener.Server.IsBound)
-                {
-                    TcpClient client = listener.AcceptTcpClient();
-                    SSLClient connection = new SSLClient(client, serverCert, VerifyClients, this);
-                }
-                connected++;
-            });
-            listenerThrewad.Start();
-        }
-
-
-
-
-
-
-
-    }
-
     /// <summary>
     /// Represents connected client 
     /// </summary>
     public sealed class SSLClient
     {
+        #region enums
+        private enum SteerCodes
+        {
+            SendText = 1,
+            SendFile = 2,
+            SendRawBytes = 3,
+            SendDirectory = 4,
+            SendDirectoryV2 = 5,
+            Confirmation = 200,
+            SendDisconnect = 99
+        }
 
+        #endregion
+
+        #region Settable fields
 
         internal bool Busy = false;
 
@@ -344,13 +52,14 @@ namespace EasySslStream.Connection.Client
         TcpClient client_ = null;
         SslStream sslstream_ = null;
 
-        //////////////////////////////////////////
-        /// Connection menagement
-        ///  <summary>
-        /// Connection menagement
+        /// <summary>
+        /// U can choose encrypiton protocols like SslProtocols = SslProtocols.TLS11|SslProtocols , leave "null" for default configuration
         /// </summary>
-        /// 
+        public SslProtocols SslProtocols;
 
+        #endregion
+
+        #region Client instance related
         internal void Stop()
         {
             sslstream_.Dispose();
@@ -405,7 +114,15 @@ namespace EasySslStream.Connection.Client
                 }
             }
         }
-        ////////////////////////////////////
+        #endregion
+
+        #region Connection Statistics
+
+        public IFileReceiveEventAndStats FileReceiveEventAndStats = ConnectionCommons.CreateFileReceive();
+        public IFileSendEventAndStats FileSendEventAndStats = ConnectionCommons.CreateFileSend();
+        public IDirectorySendEventAndStats DirectorySendEventAndStats = ConnectionCommons.CreateDirectorySendEventAndStats();
+        public IDirectoryReceiveEventAndStats DirectoryReceiveEventAndStats = ConnectionCommons.CreateDirectoryReceiveEventAndStats();
+        #endregion
 
 
         private bool ValidadeClientCert(object sender, X509Certificate certificate, X509Chain chain, SslPolicyErrors sslPolicyErrors)
@@ -435,33 +152,9 @@ namespace EasySslStream.Connection.Client
 
         }
 
-        private enum SteerCodes
-        {
-            SendText = 1,
-            SendFile = 2,
-            SendRawBytes = 3,
-            SendDirectory = 4,
-            SendDirectoryV2 = 5,
 
-            Confirmation = 200,
+        #region Connection handling
 
-
-            SendDisconnect = 99
-        }
-
-
-        /// <summary>
-        /// U can choose encrypiton protocols like SslProtocols = SslProtocols.TLS11|SslProtocols , leave "null" for default configuration
-        /// </summary>
-        public SslProtocols SslProtocols;
-
-
-
-
-        public IFileReceiveEventAndStats FileReceiveEventAndStats = ConnectionCommons.CreateFileReceive();
-        public IFileSendEventAndStats FileSendEventAndStats = ConnectionCommons.CreateFileSend();
-        public IDirectorySendEventAndStats DirectorySendEventAndStats = ConnectionCommons.CreateDirectorySendEventAndStats();
-        public IDirectoryReceiveEventAndStats DirectoryReceiveEventAndStats = ConnectionCommons.CreateDirectoryReceiveEventAndStats();
         /// <summary>
         /// Creates client instance
         /// </summary>
@@ -471,37 +164,39 @@ namespace EasySslStream.Connection.Client
         /// <param name="srvinstance"></param>
         public SSLClient(TcpClient client, X509Certificate2 serverCert, bool VerifyClients, Server? srvinstance = null)
         {
-
-
             Busy = false;
             client_ = client;
             srv = srvinstance;
+            bool cancelConnection = false;
 
             srv.ConnectedClients.Add(this);
             srv.ConnectedClientsByNumber.TryAdd(srv.ConnectedClients.Count, this);
-
-
-
-
             srv.ConnectedClientsByEndPoint.TryAdd((IPEndPoint)client.Client.RemoteEndPoint, this);
-            // Console.WriteLine("?: "+srv.ConnectedClients.Count);
 
             if (VerifyClients == false)
             {
                 sslstream_ = new SslStream(client.GetStream(), false);
                 if (this.SslProtocols == null) { sslstream_.AuthenticateAsServer(serverCert, clientCertificateRequired: false, true); }
                 else { sslstream_.AuthenticateAsServer(serverCert, clientCertificateRequired: false, this.SslProtocols, true); }
-
             }
             else
             {
                 sslstream_ = new SslStream(client.GetStream(), false, new RemoteCertificateValidationCallback(ValidadeClientCert));
                 if (this.SslProtocols == null) { sslstream_.AuthenticateAsServer(serverCert, clientCertificateRequired: true, true); }
                 else { sslstream_.AuthenticateAsServer(serverCert, clientCertificateRequired: true, this.SslProtocols, true); }
-
-
             }
 
+            
+
+            
+
+            
+
+            
+
+        }
+        private void StartServerSender()
+        {
             Thread ServerSender = new Thread(() =>
             {
                 Task.Run(async () =>
@@ -517,7 +212,6 @@ namespace EasySslStream.Connection.Client
                             w.Invoke();
                             Busy = false; privateBusy = false;
                             w = null;
-
                         }
                     }
                     catch (System.ObjectDisposedException e)
@@ -541,9 +235,9 @@ namespace EasySslStream.Connection.Client
                 }).ConfigureAwait(false).GetAwaiter().GetResult();
             });
             ServerSender.Start();
-
-            bool cancelConnection = false;
-
+        }
+        private void StartServerReceiver()
+        {
             Task.Run(async () =>
             {
                 try
@@ -612,8 +306,15 @@ namespace EasySslStream.Connection.Client
                     throw new Exceptions.ServerException($"Unknown Server Excetion: {e.Message}\n {e.StackTrace}");
                 }
             }).GetAwaiter().GetResult();
-
         }
+        #endregion
+
+
+
+
+
+
+
         private async Task<int> ConnSteer()
         {
             byte[] buffer = new byte[64];
@@ -1284,18 +985,12 @@ namespace EasySslStream.Connection.Client
                             while (sent != fs.Length)
                             {
                                 sent += fs.Read(chunk, 0, chunk.Length);
-
                                 sslstream_.Write(chunk);
                                 DirectorySendEventAndStats.CurrentSendFileCurrentBytes = sent;
-
                             }
-
                             sslstream_.Flush();
                             fs.Dispose();
                             DirectorySendEventAndStats.RaiseOnFileFromDirectorySendProcessed();
-
-                            // Task.Delay(100).Wait();
-
                         }
                         catch (System.UnauthorizedAccessException e)
                         {
@@ -1479,13 +1174,9 @@ namespace EasySslStream.Connection.Client
                                 DirectorySendEventAndStats.CurrentSendFileCurrentBytes = sent;
 
                             }
-
                             sslstream_.Flush();
                             fs.Dispose();
                             DirectorySendEventAndStats.RaiseOnFileFromDirectorySendProcessed();
-
-                            // Task.Delay(100).Wait();
-
                         }
                         catch (System.UnauthorizedAccessException e)
                         {
@@ -1503,54 +1194,21 @@ namespace EasySslStream.Connection.Client
                             {
                                 sslstream_.Write(BitConverter.GetBytes((long)-10));
                             }
-
                         }
-
-
-
-
-                        // SDCancele.Cancel();
                     }; await ServerSendingQueue.Writer.WaitToWriteAsync(); await ServerSendingQueue.Writer.WriteAsync(SendInnerDirectory);
 
                     if (LoopCancel == true)
                     {
                         break;
                     }
-
-
-
-                    //await Task.Delay(2000);
-
                     DirectorySendEventAndStats.RaiseOnFileFromDirectorySendProcessed();
                 }
-
-
                 while (ServerSendingQueue.Reader.Count > 0)
                 {
                     await Task.Delay(10);
                 }
                 SDCancele.Cancel();
-
-
             });
         }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
     }
-
-
 }
