+++ EasySslStream/Connection/ConnectionCommons.cs
@@ -1,4 +1,4 @@
-﻿namespace EasySslStream.Connection.Full
+﻿namespace EasySslStream.Connection.Client
 {
     public partial class ConnectionCommons : IFileReceiveEventAndStats, IFileSendEventAndStats, IDirectorySendEventAndStats
     {
