+++ EasySslStream/Exceptions/CertgenExceptions/CACertgenFailedException.cs
