+++ EasySslStream/Exceptions/CertgenExceptions/CSRgenFailedException.cs
