+++ EasySslStream/Exceptions/CertgenExceptions/PFXConvertException.cs
