+++ EasySslStream/Exceptions/CertgenExceptions/ServerException.cs
