+++ EasySslStream/Exceptions/CertgenExceptions/SslCertgenModeNotSetException.cs
